requirements.txt
python-telegram-bot==20.7
fastapi==0.109.0
uvicorn==0.25.0
pydantic==2.5.3
aiosqlite==0.19.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
